# *: use Equal for more efficient equality comparisons