--- a/internal/keyspan/span.go
+++ b/internal/keyspan/span.go
@@ -83,9 +83,9 @@ func (k Key) Kind() base.InternalKeyKind {
 // Equal returns true if this Key is equal to the given key. Two keys are said
 // to be equal if the two Keys have equal trailers, suffix and value. Suffix
 // comparison uses the provided base.Compare func. Value comparison is bytewise.
-func (k Key) Equal(cmp base.Compare, b Key) bool {
+func (k Key) Equal(equal base.Equal, b Key) bool {
 	return k.Trailer == b.Trailer &&
-		cmp(k.Suffix, b.Suffix) == 0 &&
+		equal(k.Suffix, b.Suffix) &&
 		bytes.Equal(k.Value, b.Value)
 }