--- a/table_stats_test.go
+++ b/table_stats_test.go
@@ -204,7 +204,7 @@ func TestTableRangeDeletionIter(t *testing.T) {
 				return err.Error()
 			}
 			defer r.Close()
-			iter, err := newCombinedDeletionKeyspanIter(cmp, r, m)
+			iter, err := newCombinedDeletionKeyspanIter(base.DefaultComparer, r, m)
 			if err != nil {
 				return err.Error()
 			}